Begin to support abstract types. start with the securitycontentobject and detection object
